moved pages into pages folder
