+++ apps/website/src/app/App.tsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { BackendPage } from './components/backend/BackendPage';
-import { EggPage } from './components/egg/EggPage';
+import { EggPage } from './components/pages/egg/EggPage';
 
-import { OverviewPage } from './components/overview/OverviewPage';
-import { ProjectsPage } from './components/projects/ProjectsPage';
-import { WorkPage } from './components/work/WorkPage';
+import { OverviewPage } from './components/pages/overview/OverviewPage';
+import { ProjectsPage } from './components/pages/projects/ProjectsPage';
+import { WorkPage } from './components/pages/work/WorkPage';
 import { urls } from './util/routes';
 
 export function App() {
