--- apps/website/src/app/components/overview/OverviewPage.tsx
@@ -1,9 +1,0 @@
-import { Page } from '../common/Page';
-import { IncompletePage } from '../common/IncompletePage';
-import { AppPaper } from '../base/AppPaper';
-
-export function OverviewPage() {
-    return <Page title="Overview">
-        <AppPaper></AppPaper>
-    </Page>;
-}
