+++ apps/website/src/app/components/pages/egg/EggPage.tsx
@@ -1,4 +1,4 @@
-import { Page } from '../common/Page';
+import { Page } from '../../common/Page';
 
 export function EggPage() {
     return <Page title="Egg!">Hello</Page>;
