+++ apps/website/src/app/components/pages/overview/OverviewPage.tsx
@@ -1,0 +1,9 @@
+import { Page } from '../../common/Page';
+import { IncompletePage } from '../../common/IncompletePage';
+import { AppPaper } from '../../base/AppPaper';
+
+export function OverviewPage() {
+    return <Page title="Overview">
+        <AppPaper></AppPaper>
+    </Page>;
+}
