+++ apps/website/src/app/components/pages/projects/base/ProjectRating.tsx
