+++ apps/website/src/app/components/pages/work/WorkPage.tsx
@@ -1,7 +1,7 @@
 import { Stack } from '@mui/material';
 import { useState } from 'react';
 
-import { Page } from '../common/Page';
+import { Page } from '../../common/Page';
 import { Work } from './Work';
 import { useWork } from './workStore';
